backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
psycopg2-binary==2.9.9
alembic==1.13.1
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[flask]==1.40.6
python-json-logger==2.0.7
celery==5.3.4
flask-mail==0.9.1
Flask-Compress==1.14
Flask-Caching==2.1.0
cryptography==41.0.8
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
flask-socketio>=5.3.0
python-engineio>=4.8.0
python-socketio>=5.10.0
requests>=2.31.0
ffmpeg-python==0.2.0